Home module

+++ src/app/home/components/project-preview/project-preview.component.ts
@@ -1,3 +1,7 @@
+import { ConfirmDialogComponent } from './../../../shared/components/confirm-dialog/confirm-dialog.component';
+import { ProjectCreateUpdateModalComponent } from './../../../shared/components/project-create-update-modal/project-create-update-modal.component';
+import { MatDialog } from '@angular/material/dialog';
+import { ICreateEditProject, ConfirmDialogModel } from './../../../core/models/dialog.model';
 import { switchMap, map } from 'rxjs';
 import { IBoard } from './../../../core/models/data.model';
 import { ProjectsDataService } from './../../../core/services/projects-data.service';
@@ -12,13 +16,48 @@ export class ProjectPreviewComponent {
 
   @Input() public project!: IBoard;
 
-  constructor(private projectsService: ProjectsDataService) {
-  
-  }
+  constructor(private projectsService: ProjectsDataService,  private projectModal: MatDialog) { }
     
   deleteProject(projectId: string) {
-    this.projectsService.deleteProject(projectId).pipe(
-      switchMap(()=>this.projectsService.getProjects().pipe(map((value)=>value)))
-    ).subscribe();
+    const dialogData = new ConfirmDialogModel(
+      'Project-modal-delete-title',
+      'Project-modal-delete-message',
+      'Delete'
+    );
+
+    const dialogRef = this.projectModal.open(ConfirmDialogComponent, {
+      maxWidth: '600px',
+      data: dialogData,
+    });
+
+    dialogRef.afterClosed().subscribe((dialogResult) => {
+      if (dialogResult) {
+        this.projectsService.deleteProject(projectId).pipe(
+          switchMap(()=>this.projectsService.getProjects().pipe(map((value)=>value)))
+        ).subscribe();
+      }
+    });
+  }
+
+  updateProject(projectId: string) {
+    const dialogData: ICreateEditProject = {
+      title:'Project-modal-edit-title',
+      projectTitleLabel:'Project-modal-title',
+      projectDescriptionLabel:'Project-modal-description',
+      commandName:'Project-modal-edit',
+    };
+
+    const dialogRef = this.projectModal.open(ProjectCreateUpdateModalComponent, {
+      maxWidth: '600px',
+      data: dialogData,
+    });
+
+    dialogRef.afterClosed().subscribe((dialogResult) => {
+      if (dialogResult) {
+        this.projectsService.updateProject(projectId, dialogResult[0], dialogResult[1]).pipe(
+          switchMap(()=>this.projectsService.getProjects().pipe(map((value)=>value)))
+        ).subscribe();
+      }
+    });
   }
 }
